scripts/requirements.txt
# Demo Zenoh publisher (protobuf RobotCommand)
eclipse-zenoh>=1.0.0
protobuf>=3.19.0
